+++ chestnut/src/main/java/com/banking/chestnut/depositRepositories/DepositRepository.java
@@ -8,6 +8,6 @@ import java.util.Optional;
 import java.util.Set;
 
 @Repository
-public interface DepositRepository extends CrudRepository<Deposit, Long> {
-    Optional<Set<Deposit>> findAllByAccountId(Long id);
+public interface DepositRepository extends CrudRepository<Deposit, Integer> {
+    Optional<Set<Deposit>> findAllByAccountId(Integer id);
 }
